Merge debugwindow+ into master to fix issues with PR #13
* Update DebugWindow with automatic text wrap
* Remove 'somefile.txt' at the end of test_debugwindow.py
* Restore previous stdout/stderr on exit
After closing the debugwindow, calls to print raised a TclError because
the Text widget had been destroyed. This commit fixes this by restoring
the previous stdout/stderr when the debugwindow is closed. An example
is added in the examples folder.
* Add flush method
The flush method is sometimes called on closing, which raised an error
because DebugWindow had no such method.
* Replace custom text wrap with tk.Text default